Bugfix for egui-miniquad: Nearest neighbour filter mode
as of egui-miniquad 0.14.0 there is a bug that makes it so that the TextureOptions are ignored by the renderer. In the case of this emulator that meant no nearest neighbour sampling for the tileset viewer, which looked pretty bad.

For now I've just cloned egui-miniquad & egui-macroquad and am building them from source with my patch added on.